+++ NEWS.md
@@ -1,9 +1,18 @@
+# mlr3proba 0.2.1
+
+* Removed `PredictionRegr` causing masking issues with `{mlr3}`
+* Bug fix in `PipeOpDistrCompositor` causing some `cdf` predictions to be lost
+* Internal fixes for `mlr3pipelines`: public train and predict methods to private
+* Added four datasets and tasks: `grace`, `actg`, `gbcs`, `whas`
+* Add `overwrite` to `crankcompositor` pipeop and pipeline
+* Bug fix in `surv.kaplan` `crank` prediction
+
 # mlr3proba 0.2.0
 
 ### Added Functionality
 
 * `MeasureSurvCindex` added. Generalises all c-index measures with a fast C++ implementation
-* Akritas estimator added to [mlr3learners.proba repo](https://www.github.com/mlr3learners/mlr3learners.proba)
+* Akritas estimator added to `mlr3learners/mlr3learners.proba`
 * Added scoring rule `MeasureSurvSchmid`
 * Addd calibration measures `MeasureSurvCalibrationBeta` and `MeasureSurvCalibrationAlpha`
 * `surv.brier` alias added for `surv.graf`
@@ -25,7 +34,7 @@
 
 ### Moved Functionality
 
-- `LearnerSurvGlmnet`, `LearnerSurvCVGlmnet`, `LearnerSurvXgboost` and `LearnerSurvRanger` have been moved to the [mlr3learners repo](https://www.github.com/mlr-org/mlr3learners)
+- `LearnerSurvGlmnet`, `LearnerSurvCVGlmnet`, `LearnerSurvXgboost` and `LearnerSurvRanger` have been moved to `mlr-org/mlr3learners`
 
 - `LearnerSurvGBM` has been moved to https://www.github.com/mlr3learners/mlr3learners.gbm